Changed: Minors server changes from Hg

+++ src/entities_game_service/guild_manager/guild.cpp
@@ -698,7 +698,6 @@ void CGuild::unregisterGuild()
 //
 //}
 
-
 //----------------------------------------------------------------------------
 void CGuild::removeMission(CMissionGuild * mission, TMissionResult result)
 {
@@ -798,7 +797,6 @@ void CGuild::putItem( CCharacter * user, uint32 slot, uint32 quantity, uint16 se
 		CCharacter::sendDynamicSystemMessage( user->getId(),"GUILD_ITEM_CANT_BE_PUT" );
 		return;
 	}
-
 
 	// try to move the required quantity of the item
 	if ( CInventoryBase::moveItem(
@@ -1481,7 +1479,7 @@ void CGuild::addOwnedOutpost(TAIAlias outpostAlias)
 
 		// update all outposts following the new one (included)
 		// it will shift right
-		uint i = _OwnedOutposts.size() - 1;
+		uint i = (uint)_OwnedOutposts.size() - 1;
 		while (i < _OwnedOutposts.size() + _ChallengedOutposts.size())
 		{
 			if (!updateOutpostDB(i))
@@ -1572,7 +1570,7 @@ void CGuild::addChallengedOutpost(TAIAlias outpostAlias)
 		nlassert(getOutpostCount() < OUTPOSTENUMS::MAX_OUTPOST);
 
 		_ChallengedOutposts.push_back(outpostAlias);
-		updateOutpostDB( _OwnedOutposts.size()+_ChallengedOutposts.size()-1 );
+		updateOutpostDB( (uint32)(_OwnedOutposts.size()+_ChallengedOutposts.size()-1) );
 
 		updateGUILD_OUTPOST_CANDEL();
 	}
@@ -1597,7 +1595,7 @@ void CGuild::removeChallengedOutpost(TAIAlias outpostAlias)
 
 	// update all outposts following the removed one
 	// it will shift left
-	i += _OwnedOutposts.size();
+	i += (uint)_OwnedOutposts.size();
 	while (i < _OwnedOutposts.size() + _ChallengedOutposts.size())
 	{
 		if (!updateOutpostDB(i))
@@ -1618,7 +1616,7 @@ void CGuild::removeChallengedOutpost(TAIAlias outpostAlias)
 //-----------------------------------------------------------------------------
 uint32 CGuild::getOutpostCount() const
 {
-	return (_OwnedOutposts.size() + _ChallengedOutposts.size());
+	return (uint32)(_OwnedOutposts.size() + _ChallengedOutposts.size());
 }
 
 //-----------------------------------------------------------------------------
@@ -1711,7 +1709,7 @@ bool CGuild::getOutpostDBIndex(TAIAlias outpostAlias, uint32 & outpostIndex, boo
 	{
 		if (_ChallengedOutposts[i] == outpostAlias)
 		{
-			outpostIndex = _OwnedOutposts.size() + i;
+			outpostIndex = (uint32)_OwnedOutposts.size() + i;
 			ownedOutpost = false;
 			return true;
 		}
@@ -1734,7 +1732,7 @@ bool CGuild::updateOutpostDB(uint32 outpostIndex)
 	}
 	else
 	{
-		i -= _OwnedOutposts.size();
+		i -= (uint32)_OwnedOutposts.size();
 		if (i < _ChallengedOutposts.size())
 		{
 			outpostAlias = _ChallengedOutposts[i];
@@ -2033,7 +2031,7 @@ void	IGuild::updateMembersStringIds()
 //-----------------------------------------------------------------------------
 /**
  * This class is used to load old guild inventory, DO NOT BREAK IT!
- * \author Sébastien 'kxu' Guignot
+ * \author Sebastien 'kxu' Guignot
  * \author Nevrax France
  * \date 2005
  */
